Move types around
